+++ src/main/java/seedu/duke/Duke.java
@@ -8,13 +8,13 @@ import commands.ViewExpenseCommand;
 import exceptions.BudgetTrackerException;
 import expenses.Ui;
 import income.IncomeParser;
+import savings.RunSavings;
 import summary.Summary;
 import summary.ui.SummaryDisplay;
 import ui.HelpDisplay;
 import expenses.BudgetTracker;
 import expenses.ExpenseParser;
 import expenses.ExpenseList;
-import savings.Saving;
 import alerts.FundsAlert;
 import alerts.AlertParser;
 import java.util.Scanner;
@@ -27,7 +27,6 @@ public class Duke {
     private final Ui ui;
     private final BudgetTracker tracker;
     private final ExpenseList expenseList;
-    private final Saving saving;
     private final FundsAlert fundsAlert;
 
     /**
@@ -41,7 +40,6 @@ public class Duke {
         ui = new Ui();
         tracker = new BudgetTracker();
         expenseList = new ExpenseList();
-        saving = new Saving(summary);
         fundsAlert = new FundsAlert(ui);
         summary.registerObserver(fundsAlert);
     }
@@ -147,7 +145,7 @@ public class Duke {
 
                 // Handle savings commands
                 if (fullCommand.contains("savings")) {
-                    saving.run(fullCommand);
+                    RunSavings.run(fullCommand);
                     commandRecognized = true;
                 }
